backend/requirements.txt
# Flask and extensions
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0

# Database
pymysql==1.1.0
PyMySQL==1.1.0
cryptography==41.0.3
pymongo==4.5.0

# Image processing
Pillow==10.0.0

# Utils
python-dotenv==1.0.0
Werkzeug==2.3.7
bson==0.5.10
openai==0.28.0
requests==2.31.0

# Deployment
gunicorn==21.2.0
